Editing: 27 task, 15 var (2016) - int to long (to parse 500k+)

+++ evich_2016/15/task_27.cpp
@@ -3,7 +3,7 @@
 int main() {
     const int PRICE_LIMIT = 240,
               PROPOSAL_NAME_LIMIT = 4;
-    int N; // by condition: the number of proposals
+    long N; // by condition: the number of proposals
     std::cin >> N;
     if (N <= 0) {
         return -1;
